target_pipeline/requirements.txt
sqlalchemy>=2.0
python-dotenv>=1.0
pytest>=8.0
psycopg[binary]>=3.1
